refactor: update layout classes for chapter preview and rename draft sidebar to content sidebar
