+++ src/frontend/shared/content-sidebar.vue
